+++ packages/stream_video/lib/stream_video.dart
@@ -55,6 +55,7 @@ export 'src/sfu/data/models/sfu_participant_source.dart';
 export 'src/sfu/data/models/sfu_track_type.dart';
 export 'src/sorting/call_participant_sorting_presets.dart';
 export 'src/sorting/call_participant_state_sorting.dart';
+export 'src/sorting/sort_participants.dart';
 export 'src/stream_video.dart';
 export 'src/token/token.dart';
 export 'src/types/other.dart';